@@ -4,8 +4,8 @@ namespace SvaSorcery.Patterns.Enterprise.ORM.IdentityMap.Types
 {
     public class GenericIdentityMap
     {
-        protected static readonly GenericIdentityMap _instance = new GenericIdentityMap();
-        protected readonly Dictionary<int, object> _pool = new Dictionary<int, object>();
+        protected static readonly GenericIdentityMap _instance = new();
+        protected readonly Dictionary<int, object> _pool = new();
 
         public static T Get<T>(int id) where T : DomainObject
             => (T)_instance._pool[id];
