@@ -7,29 +7,27 @@ namespace SvaSorcery.Patterns.Enterprise.Base.Plugin
 {
     public class HttpContextUserPersonPlugin : IPersonExtensions
     {
-        private HttpContext _httpContext;
+        private readonly HttpContext _httpContext;
 
         public HttpContextUserPersonPlugin(IHttpContextAccessor httpContextAccessor)
         {
             _httpContext = httpContextAccessor.HttpContext;
         }
 
-        public Person RepositoryOwner()
-            => new Person()
-            {
-                Id = int.Parse(_httpContext.User.Claims.FirstOrDefault(x => x.Type == "id")?.Value),
-                FirstName = _httpContext.User.Identity.Name.Split(' ')[0],
-                LastName = _httpContext.User.Identity.Name.Split(' ')[1],
-                Email = _httpContext.User.Claims.FirstOrDefault(x => x.Type == "email")?.Value
-            };
+        public Person RepositoryOwner() => new()
+        {
+            Id = int.Parse(_httpContext.User.Claims.FirstOrDefault(x => x.Type == "id")?.Value),
+            FirstName = _httpContext.User.Identity.Name.Split(' ')[0],
+            LastName = _httpContext.User.Identity.Name.Split(' ')[1],
+            Email = _httpContext.User.Claims.FirstOrDefault(x => x.Type == "email")?.Value
+        };
 
-        public Person NullPerson()
-            => new Person()
-            {
-                Id = 0,
-                FirstName = null,
-                LastName = null,
-                Email = null
-            };
+        public Person NullPerson() => new()
+        {
+            Id = 0,
+            FirstName = null,
+            LastName = null,
+            Email = null
+        };
     }
 }
