@@ -5,8 +5,6 @@ namespace SvaSorcery.Patterns.Enterprise.Base.ServiceStub
     public static class Extensions
     {
         public static void AddCalculatorService(this IServiceCollection services)
-        {
-            services.AddScoped<ICalculatorService, CalculatorServiceStub>();
-        }
+            => services.AddScoped<ICalculatorService, CalculatorServiceStub>();
     }
 }
