@@ -2,10 +2,10 @@
 
 namespace SvaSorcery.Patterns.Enterprise.Base.ValueObject
 {
-    public class DateRangeValueObject
+    public record DateRangeValueObject
     {
-        public DateTime From { get; set; }
-        public DateTime To { get; set; }
+        public DateTime From { get; private set; }
+        public DateTime To { get; private set; }
 
         public DateRangeValueObject(DateTime from, DateTime to)
         {
@@ -15,9 +15,13 @@ namespace SvaSorcery.Patterns.Enterprise.Base.ValueObject
 
         public DateRangeValueObject(string fromString, string toString)
         {
-            TryParse(fromString, out DateTime from);
-            TryParse(toString, out DateTime to);
-            new DateRangeValueObject(from, to);
+            if (!TryParse(fromString, out DateTime from))
+                throw new ArgumentException(nameof(fromString));
+
+            if (!TryParse(toString, out DateTime to))
+                throw new ArgumentException(nameof(toString));
+
+            _ = new DateRangeValueObject(from, to);
         }
 
         public bool IsBetween(DateTime dateToCheck)
@@ -27,7 +31,7 @@ namespace SvaSorcery.Patterns.Enterprise.Base.ValueObject
 
         public bool IsFinished => To < DateTime.Now;
 
-        private bool TryParse(string dateString, out DateTime date)
+        private static bool TryParse(string dateString, out DateTime date)
             => DateTime.TryParse(dateString, out date);
     }
 }
