@@ -4,17 +4,17 @@ namespace SvaSorcery.Patterns.Enterprise.Domain.TransactionScript
 {
     public class Money
     {
-        public decimal Amount { get; private set; }
+        public decimal Amount { get; }
         public Currency Currency { get; }
 
         public Money(decimal amount, Currency currency)
-        {
-            Currency = currency;
-            Amount = amount * currency.CentFactor;
-        }
+            => (Amount, Currency) = (amount * currency.CentFactor, currency);
+
+        public Money NewMoney(decimal amount)
+            => new(amount, Currency);
 
         public Money Add(Money value)
-            => new Money(Amount += value.Amount, Currency);
+            => NewMoney(Amount + value.Amount);
 
         public Money[] Allocate(int parts)
         {
@@ -27,11 +27,14 @@ namespace SvaSorcery.Patterns.Enterprise.Domain.TransactionScript
             return result;
         }
 
-        public static Money Dollars(decimal amount) => new Money(amount, new Currency("USD", 2));
-        public static Money Roubles(decimal amount) => new Money(amount, new Currency("RUR", 2));
+        public static Money Dollars(decimal amount)
+            => new(amount, new Currency("USD", 2));
+
+        public static Money Roubles(decimal amount)
+            => new(amount, new Currency("RUR", 2));
     }
 
-    public class Currency
+    public record Currency
     {
         public string Name { get; }
         public int FractionDigits { get; }
