@@ -5,16 +5,15 @@ namespace SvaSorcery.Patterns.Enterprise.InputOutput.UpdateFactory
 {
     public class HospitalScheduleSlotUpdateFactory : IUpdateFactory<HospitalScheduleSlot>
     {
-        public Hashtable NewUpdate(HospitalScheduleSlot model)
-            => new Hashtable
-            {
-                { nameof(model.Date), model.Date },
-                { nameof(model.TimeFrom), model.TimeFrom },
-                { nameof(model.TimeTo), model.TimeTo },
-                { nameof(model.SpecId), model.SpecId },
-                { nameof(model.PatientId), model.PatientId },
-                { nameof(model.DoctorId), model.DoctorId },
-                { nameof(model.RoomId), model.RoomId }
-            };
+        public Hashtable NewUpdate(HospitalScheduleSlot model) => new()
+        {
+            { nameof(model.Date), model.Date },
+            { nameof(model.TimeFrom), model.TimeFrom },
+            { nameof(model.TimeTo), model.TimeTo },
+            { nameof(model.SpecId), model.SpecId },
+            { nameof(model.PatientId), model.PatientId },
+            { nameof(model.DoctorId), model.DoctorId },
+            { nameof(model.RoomId), model.RoomId }
+        };
     }
 }
