@@ -2,16 +2,5 @@
 
 namespace SvaSorcery.Patterns.Enterprise.ORM.IdentityMap
 {
-    public class AlbumIdentity : DomainObject
-    {
-        public string Title { get; }
-        public string Artist { get; }
-
-        public AlbumIdentity(int id, string title, string artist)
-        {
-            Id = id;
-            Title = title;
-            Artist = artist;
-        }
-    }
+    public record AlbumIdentity(int Id, string Title, string Artist) : DomainObject(Id);
 }
