@@ -7,13 +7,13 @@ namespace SvaSorcery.Patterns.Enterprise.Presentation.AppController.Services
 {
     public class PersonsRepository : IRepository<Person>
     {
-        private readonly HashSet<Person> _personsDatabaseTable = new HashSet<Person>
+        private readonly HashSet<Person> _personsDatabaseTable = new()
         {
-            new Person() { Id = 1, FirstName = "Alice", LastName = "Sender", Email = "alice@personcached.io" },
-            new Person() { Id = 2, FirstName = "Bob", LastName = "Receipient", Email = "bob@personcached.io" },
-            new Person() { Id = 3, FirstName = "Carol", LastName = "Middleman", Email = "carol@personcached.io" },
-            new Person() { Id = 4, FirstName = "Eve", LastName = "Eavesdropper", Email = "eve@personcached.io" },
-            new Person() { Id = 5, FirstName = "Mallory", LastName = "Malicious", Email = "mallory@personcached.io" },
+            new() { Id = 1, FirstName = "Alice", LastName = "Sender", Email = "alice@personcached.io" },
+            new() { Id = 2, FirstName = "Bob", LastName = "Receipient", Email = "bob@personcached.io" },
+            new() { Id = 3, FirstName = "Carol", LastName = "Middleman", Email = "carol@personcached.io" },
+            new() { Id = 4, FirstName = "Eve", LastName = "Eavesdropper", Email = "eve@personcached.io" },
+            new() { Id = 5, FirstName = "Mallory", LastName = "Malicious", Email = "mallory@personcached.io" },
         };
 
         public void CreateOrUpdate(Person item) => _personsDatabaseTable.Add(item);
