@@ -2,17 +2,5 @@
 
 namespace SvaSorcery.Patterns.Enterprise.Presentation.Common.Commands.Commands
 {
-    public class CalculateRevenueRecognitions : ICommand
-    {
-        public long ContractId { get; }
-
-        protected CalculateRevenueRecognitions()
-        {
-        }
-
-        public CalculateRevenueRecognitions(long contractId)
-        {
-            ContractId = contractId;
-        }
-    }
+    public record CalculateRevenueRecognitions(long ContractId) : ICommand;
 }
