@@ -4,21 +4,5 @@ using SvaSorcery.Patterns.Enterprise.Presentation.Common.Commands.Types;
 
 namespace SvaSorcery.Patterns.Enterprise.Presentation.Common.Commands.Commands
 {
-    public class CreateRecognition : ICommand
-    {
-        public long ContractId { get; }
-        public DateTime RecognizedAt { get; }
-        public Money Money { get; set; }
-
-        protected CreateRecognition()
-        {
-        }
-
-        public CreateRecognition(long contractId, DateTime recognizedAt, Money money)
-        {
-            ContractId = contractId;
-            RecognizedAt = recognizedAt;
-            Money = money;
-        }
-    }
+    public record CreateRecognition(long ContractId, DateTime RecognizedAt, Money Money) : ICommand;
 }
