@@ -5,19 +5,5 @@ using SvaSorcery.Patterns.Enterprise.Presentation.Common.Commands.Types;
 
 namespace SvaSorcery.Patterns.Enterprise.Presentation.Common.Commands.Queries
 {
-    public class FindRecognitions : IQuery<IEnumerable<Money>>
-    {
-        public long ContractId { get; }
-        public DateTime RecognizedAt { get; }
-
-        protected FindRecognitions()
-        {
-        }
-
-        public FindRecognitions(long contractId, DateTime recognizedAt)
-        {
-            ContractId = contractId;
-            RecognizedAt = recognizedAt;
-        }
-    }
+    public record FindRecognitions(long ContractId, DateTime RecognizedAt) : IQuery<IEnumerable<Money>>;
 }
