@@ -4,19 +4,5 @@ using SvaSorcery.Patterns.Enterprise.Presentation.Common.Commands.Types;
 
 namespace SvaSorcery.Patterns.Enterprise.Presentation.Common.Commands.Queries
 {
-    public class GetRecognizedRevenue : IQuery<Money>
-    {
-        public long ContractId { get; }
-        public DateTime RecognizedAt { get; }
-
-        protected GetRecognizedRevenue()
-        {
-        }
-
-        public GetRecognizedRevenue(long contractId, DateTime recognizedAt)
-        {
-            ContractId = contractId;
-            RecognizedAt = recognizedAt;
-        }
-    }
+    public record GetRecognizedRevenue(long ContractId, DateTime RecognizedAt) : IQuery<Money>;
 }
