@@ -1,22 +1,4 @@
 ﻿namespace SvaSorcery.Patterns.Enterprise.Presentation.Common.Views.Models
 {
-    public class Person
-    {
-        public int Id { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Email { get; set; }
-
-        protected Person()
-        {
-        }
-
-        public Person(int id, string firstName, string lastName, string email)
-        {
-            Id = id;
-            FirstName = firstName;
-            LastName = lastName;
-            Email = email;
-        }
-    }
+    public record Person(int Id, string FirstName, string LastName, string Email);
 }
